+++ kernel/__init__.py
@@ -20,6 +20,9 @@ from kernel.marketplace_domain import (
 )
 from kernel.marketplace_store import MarketplaceStore
 from kernel.cluster import ClusterManager
+from kernel.observability import ObservabilityEngine
+from kernel.observability_domain import LogEntry, MetricRecord, MetricType, TraceSpan
+from kernel.observability_store import ObservabilityStore
 from kernel.plan_store import PlanStore
 from kernel.semantic_graph import Entity, GraphQuery, InferenceRule, KnowledgeGraph, QueryResult, Relation
 from kernel.swarm import SwarmCoordinator
@@ -54,4 +57,10 @@ __all__ = [
     "CatalogEntry",
     "NodeInfo",
     "ClusterTopology",
+    "ObservabilityEngine",
+    "ObservabilityStore",
+    "MetricRecord",
+    "TraceSpan",
+    "LogEntry",
+    "MetricType",
 ]